Arreglar carga de archivo de configuración y gráficas dinámicas
