+++ config-file.csv
@@ -1,3 +1,4 @@
 ﻿What and how to plot;Graph names;Grid;Rounded corners;Graph and table space;Data resolution;Tab;Popup
 2,1|3,2|4,2|5,3;Histograma de edades|Frecuencia en preferencia de pierna|Escolaridad|Posición de juego;2x2;0|0.2|1|1;6|12|8|6;1|2|1|0;Analísis estático;1
 4,1|5,4;Diagrama 1|Diagrama 2;1x2;0|0;6|10;2|1;Frecuencias;1
+4,1|5,4;Diagrama 1|Diagrama 2;1x2;0|0;6|10;2|1;Andres;1
